+++ CHANGELOG.md
@@ -49,6 +49,10 @@
 
 ### Bug fixes and improvements
 
+- server: changes to the Rest Endpoints OpenAPI specification:
+    - the nullability of items in the output is now always correctly reported
+    - scalars other than UUID are more consistently inlined
+    - objects now have a title and, when available, the same description as in the GraphQL schema
 - server: fix dropping column from a table that has update permissions (fix #8415)
 - console: Hide TimescaleDB internal schema from data tab
 - console: support naming convention in source customization for postgres DB [CON-297]